style: renaming component files
